requirements.txt
python-telegram-bot==20.6
openai>=1.3.0
pandas>=1.5.0
PyPDF2>=3.0.0

# macOS specific
certifi>=2021.10.8
urllib3>=1.26.0